v 1.2.0

+++ tasks.py
@@ -2,73 +2,54 @@ import os
 import invoke
 import pathlib
 import src.commandcript as commandcript
-from src.commandcript import ENV_CONTEXT
 
 
-ENV_CONTEXT\
-    .add_env_var('PROJECT_GIT_DIR', f'{__file__}/../')\
-    .add_env_var('COMMANDSCRIPT_SCRIPT_DIR', f'{ENV_CONTEXT.PROJECT_GIT_DIR}/.generated')\
-    .add_env_var('PROJECT_DIST_DIR', f'{ENV_CONTEXT.PROJECT_GIT_DIR}/dist')
+commandcript.ENV_CONTEXT\
+    .add_env_var('COMMANDSCRIPT_SCRIPT_DIR', str(pathlib.Path(f'{__file__}').parent / '.generated'))\
+    .add_env_var('PROJECT_GIT_DIR', str(pathlib.Path(f'{__file__}').parent))\
+    .add_env_var('PROJECT_SRC_DIR', '${PROJECT_GIT_DIR}/src')\
+    .add_env_var('PROJECT_DIST_DIR', '${PROJECT_GIT_DIR}/dist')
 
 
 @commandcript.script_task()
 def get_info(ctx):
     """
-    Print to console information about active configuration of invoke-tasks
+    Print to console information about active configuration of commandcript-tasks
     """
-    from prettytable import PrettyTable
+    names = [value.name for value in commandcript.ENV_CONTEXT.values()]
+    hold_values = [value.hld for value in commandcript.ENV_CONTEXT.values()]
+    expanded_values = [value.exp for value in commandcript.ENV_CONTEXT.values()]
+    width = max(max(len(key) for key in names), max(len(item) for item in hold_values), max(len(item) for item in expanded_values), 25)
+    commandcript.INFO.log_line("Active environment configuration:")
+    commandcript.INFO.log_line(f"| {'Env-var name':<{width}} | {'Env-var hold-value':<{width}} | {'Env-var expanded-value':<{width}} |")
+    commandcript.INFO.log_line(f"|-{'-' * width}-|-{'-' * width}-|-{'-' * width}-|")
+    for i in range(len(names)):
+        key = names[i]
+        hold_value = hold_values[i]
+        expanded_value = expanded_values[i]
+        if expanded_value == hold_value:
+            expanded_value = '-'
+        commandcript.INFO.log_line(f"| {key:<{width}} | {hold_value:<{width}} | {expanded_value:<{width}} |")
 
-    table = PrettyTable()
-    table.align = "l"
-    table.field_names = ["ENV-name", "ENV-value"]
-    for key, value in ENV_CONTEXT.items():
-        table.add_row([key, value])
 
-    commandcript.INFO\
-        .log_line("Active environment configuration:") \
-        .log_line(f"{table}")
-
-
-@commandcript.script_task(
-    help={
-        'cwd': 'working directory for yapf python files (by default: ENV_CONTEXT["PROJECT_GIT_DIR"])',
-        'style-yapf': 'path to the .style.yapf (by default: ENV_CONTEXT["PROJECT_GIT_DIR"]/.style.yapf)',
-        'dirs': 'list of directories where python files should be discovering (by default: ["./"])',
-    },
-    iterable=['dirs'])
-def yapf(ctx, cwd: str = f'{ENV_CONTEXT.PROJECT_GIT_DIR}', style_yapf: str = f'{ENV_CONTEXT.PROJECT_GIT_DIR}/.style.yapf', dirs: list = None):
+@commandcript.script_task()
+def yapf(ctx):
     """
-    Format python files with script-tasks
+    Format python files in Fuzz
     """
-
-    def collect_file(dir):
-        files = []
-        for item in os.listdir(dir):
-            item = pathlib.Path(os.path.join(f'{dir}', item))
-            if item.is_file():
-                if item.name.endswith('.py'):
-                    files.append(f'"{item.as_posix()}"')
-            elif item.is_dir():
-                if not item.name.startswith('.'):
-                    files.extend(collect_file(f'{item.as_posix()}'))
-        return files
-
-    if not dirs:
-        dirs = ['./']
-
-    for dir in dirs:
-        log_file = 'yapf_' + '-'.join('/'.split(dir))
-        log_file = log_file.replace('.', '')
-        commandcript.ScriptExecutor(ctx.script_dir, ctx.launch)\
-            .add_cwd(cwd)\
-            .add_command([
-                    f'yapf',
-                    f'--style {style_yapf}',
-                    f'--verbose',
-                    f'--in-place',
-                    *collect_file(f'{cwd}/{dir}')
-                ])\
-            .execute(log=log_file)
+    commandcript.ScriptExecutor(ctx.script_dir, ctx.launch)\
+        .add_cwd(commandcript.ENV_CONTEXT.PROJECT_GIT_DIR)\
+        .add_command([
+                "yapf",
+                "--style .style.yapf",
+                "--verbose",
+                "--recursive",
+                "--in-place",
+                "--parallel",
+                f"--exclude '**.venv**'",
+                f"{commandcript.ENV_CONTEXT.PROJECT_SRC_DIR}",
+            ])\
+        .execute(log="yapf.log")
 
 
 @commandcript.script_task(
@@ -81,15 +62,16 @@ def prepare_build(ctx, install_build_tool: bool = False, clean_dist=False):
     Prepare build before uploading on PyPl
     """
     if clean_dist:
-        dist_dir = ENV_CONTEXT['PROJECT_DIST_DIR']
-        for filename in os.listdir(dist_dir):
-            file_path = os.path.join(dist_dir, filename)
-            if os.path.isfile(file_path):
-                os.remove(file_path)
-                commandcript.INFO.log_line(f'Deleted dist: {file_path}')
+        dist_dir = pathlib.Path(commandcript.ENV_CONTEXT.PROJECT_DIST_DIR.exp)
+        if dist_dir.exists():
+            for filename in os.listdir(dist_dir):
+                file_path = os.path.join(dist_dir, filename)
+                if os.path.isfile(file_path):
+                    os.remove(file_path)
+                    commandcript.INFO.log_line(f'Deleted dist: {file_path}')
 
     commandcript.ScriptExecutor(ctx.script_dir, ctx.launch)\
-        .add_cwd(ENV_CONTEXT['PROJECT_GIT_DIR'])\
+        .add_cwd(commandcript.ENV_CONTEXT.PROJECT_GIT_DIR.hld)\
         .add_command([f'pip install --upgrade build'] if install_build_tool else None)\
         .add_command([f'python -m build'])\
         .execute(log='prepare_build.log')
@@ -104,7 +86,7 @@ def publish_build(ctx, install_uploading_tool: bool = False, upload_on_test: boo
     """
     Uploading build on PyPl
     """
-    script = commandcript.ScriptExecutor(ctx.script_dir, ctx.launch).add_cwd(ENV_CONTEXT['PROJECT_GIT_DIR'])
+    script = commandcript.ScriptExecutor(ctx.script_dir, ctx.launch).add_cwd(commandcript.ENV_CONTEXT.PROJECT_GIT_DIR.hld)
     if install_uploading_tool:
         script.add_command([f'pip install --upgrade twine'])
     if upload_on_test:
